Memory/requirements.txt
# Core dependencies for Phase 1 WhatsApp Memory Bot

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# WhatsApp integration
requests==2.31.0

# Azure Speech Services
azure-cognitiveservices-speech==1.34.0

# Encryption
cryptography==41.0.7

# Environment and configuration
python-dotenv==1.0.0

# JSON and YAML processing
pyyaml==6.0.1

# Async support
aiofiles==23.2.1

# Logging
colorlog==6.8.0

# System monitoring
psutil==5.9.6
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Optional: Redis for session management (Phase 2)
# redis==5.0.1
coverage
pytest
pytest-asyncio
pytest-cov
ffmpeg-python
redis

# Claude AI Integration
anthropic==0.34.0